+++ code/src/solver/solver.cpp
@@ -28,6 +28,10 @@ void Solver::solve() {
         // Preprocessing
         {
             auto _prep_timer = scoped_timer("preprocess");
+
+            // IMPORTANT: KEEP THIS BEFORE EVERYTHING ELSE, SO THE EMPTY STATE ASSUMPTION HOLDS FOR EVERYTHING, INCLUDING PREPROCESSING
+            prep_initial_state_removal_();
+
             if (params_.get<cli_desc::preprocess, std::string>() != "0") {
                 logger_[INFO] << "Preprocessing instance";
                 preprocess_();